refactor: 전체 구조 수정
