+++ src/main/java/lotto/controller/LottoController.java
@@ -4,9 +4,9 @@ import lotto.domain.Lotto;
 import lotto.domain.LottoService;
 import lotto.domain.WinningLotto;
 import lotto.domain.lottomachine.RandomLottoMachine;
-import lotto.domain.primitive.LottoNumber;
-import lotto.domain.primitive.Money;
-import lotto.domain.primitive.Ticket;
+import lotto.domain.LottoNumber;
+import lotto.domain.Money;
+import lotto.domain.Ticket;
 import lotto.view.InputView;
 import lotto.view.OutputView;
