+++ src/test/java/lotto/domain/MoneyTest.java
@@ -1,5 +1,6 @@
-package lotto.domain.primitive;
+package lotto.domain;
 
+import lotto.domain.Money;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
