+++ src/test/java/lotto/domain/WinningLottoTest.java
@@ -1,8 +1,5 @@
 package lotto.domain;
 
-import lotto.domain.WinningLotto;
-import lotto.domain.Lotto;
-import lotto.domain.primitive.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
